Organizando pastas
Transferindo o arquivo token para components
