+++ components/tokenFunc.php
